+++ examples/workshop_vsecm_eso/hack/docker-build.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+
+cd app || exit
+
+docker build -t localhost:5000/vsecm-scout:v1 .