+++ examples/workshop_vsecm_eso/hack/token.go
@@ -1,0 +1,26 @@
+package main
+
+import (
+	"fmt"
+	"os"
+	"time"
+
+	"github.com/golang-jwt/jwt/v4"
+)
+
+func main() {
+	secret := []byte(os.Getenv("JWT_SECRET"))
+	claims := jwt.MapClaims{
+		"sub":  "vsecm-scout-client",
+		"name": "VSecM Scout Client",
+		"iat":  time.Now().Unix(),
+		"exp":  time.Now().Add(time.Hour * 24 * 30).Unix(),
+	}
+	token := jwt.NewWithClaims(jwt.SigningMethodHS256, claims)
+	tokenString, err := token.SignedString(secret)
+	if err != nil {
+		fmt.Println("Error creating token:", err)
+		os.Exit(1)
+	}
+	fmt.Print(tokenString)
+}